keep StateProviderFactory on BPFactory only, not BP3
